pyproject.toml
[project]
name = "news-aggregator"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "feedparser>=6.0",
    "requests>=2.31",
    "python-dateutil>=2.9",
    "pyyaml>=6.0",
]
